✏️ Fix : 로그 추가 및 테스트를 위한 일부코드 주석처리

+++ business/src/main/java/com/mzbr/business/member/dto/MemberNicknameChangeDto.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @Builder
 public class MemberNicknameChangeDto {
-	private final int userId;
+	private final long userId;
 	private final String nickname;
 
 	public static MemberNicknameChangeDto of(int userId, String nickname) {
