+++ business/src/main/java/com/mzbr/business/video/dto/VideoViewDto.java
@@ -13,7 +13,7 @@ public class VideoViewDto {
 	long memberId;
 	long videoId;
 
-	public static VideoViewDto of(long memberId, int videoId) {
+	public static VideoViewDto of(long memberId, long videoId) {
 		return VideoViewDto.builder()
 			.memberId(memberId)
 			.videoId(videoId)
